requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
bcrypt==4.3.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==45.0.4
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.13
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
limits==4.2
packaging==24.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.34.3
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
